Fix WP import so topics have the right keys
The Juvia JS looks up topics using a key that it gets from `location.pathname`. For `mydomain.com/my/post`, this is `/my/post`. When the WP import script runs, it creates the keys as `mydomain.com/my/post`, meaning that the topics are not correctly fetched by the JS on each page.